Single node config-less refactor
This is a major refactor of the pusher node. Instead of relying on a config node to hold the connection and adding a pusher node for each channel subscribtion, each pusher node now has its own connection. Multiple channel subscriptions can be set up using the input port. The goal is to declutter workflows which subscribe to several channels.

This refactor also uses a new class-based Node-RED api definition. Is it superior? Hmm.. A matter of taste.